+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 
+## 4.2.0
+August 04, 2023
+- Introduced onError(details) {} to the StateX class
+- _key = ValueKey<StateX>() with the InheritedWidget, _StateXInheritedWidget
+- List<StateXController> get controllerList is made available.
+- Introduce more documentation
+
 ## 4.1.1
 August 01, 2023
 - Unlike StateX, AppStateX uses both its InheritedWidget and its buildIn() function
   but InheritedWidget can be called separately allowing for only dependencies rebuilds 
+- Updated test scripts
+
 
 ## 4.1.0
 July 31, 2023